backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
pydantic==2.9.2
ollama==0.3.1
aiosqlite==0.20.0